feat: Add GameOver GUI & Global Reset system.
feat: Add GameOver GUI & Global Reset system.

When the player has their life reset it is now possible to see a death menu appearing. The menu has integration with the reset API present in SGameController, thus allowing the game to restart completely without the need to complete the process and exit the game to reopen it.