fix: Fix attack issues on 5th enemy.
The first few instances of the 5th enemy would normally attack the player, but after a certain period of time, all other subsequent instances of the 5th enemy would stop attacking the player.

This strange behavior was caused due to incorrect use of Tasks. Now the 5th enemy works synchronously.